+++ src/main/java/com/gym/dto/TrainingDto.java
@@ -8,11 +8,11 @@ import java.sql.Date;
 @Data
 @AllArgsConstructor
 public class TrainingDto {
-    private Integer id;
-    private Integer customerId;
-    private Integer instructorId;
+    private Long id;
+    private Long customerId;
+    private Long instructorId;
     private String trainingName;
-    private Integer trainingTypeId;
+    private Long trainingTypeId;
     private Date trainingDate;
     private Integer trainingDuration;
 }
